refactor: pages를  별도의 디렉토리로 분리한다
